requirements.txt
asgiref==3.5.1
dj-static==0.0.6
Django==4.0.4
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
Faker==13.11.1
gunicorn==20.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
